evaluation dependencies uploaded
For some evaluations (esp. comparing redcliff-s against supervised causal baselines) we need a highly specific set of dependencies which we found necessary to define in their own config file. It is recommended that this environment only be used as necessary, due to its heavy reliance on pip (over, say, conda) installations.